Rename models package to tfmodels
