+++ internal/provider/tfmodels/deployment_resource_model.go
@@ -1,4 +1,4 @@
-package models
+package tfmodels
 
 import (
 	"fmt"
